fix: bugs in selectors, less timeout, reorganise structure
