+++ .mocharc.yml
@@ -1,0 +1,10 @@
+parallel: false
+recursive: true
+color: true
+timeout: 20000
+exit: true
+require: 
+  - 'ts-node/register'
+  - 'test/global.ts'
+spec:
+  - 'test/**/*.spec.ts'
