+++ package.json
@@ -45,7 +45,7 @@
   "devDependencies": {
     "@jest/types": "^27.1.1",
     "@types/chai": "^4.2.22",
-    "@types/mocha": "^7.0.2",
+    "@types/mocha": "^9.1.1",
     "@typescript-eslint/eslint-plugin": "^4.15.0",
     "@typescript-eslint/parser": "^4.15.0",
     "chai": "^4.3.4",
@@ -55,7 +55,7 @@
     "eslint-plugin-prettier": "^3.3.1",
     "ganache": "^7.4.3",
     "jest-environment-node": "^27.1.1",
-    "mocha": "7.2.0",
+    "mocha": "^10.0.0",
     "prettier": "^2.2.1",
     "puppeteer": "14.0.0",
     "serve-handler": "5.0.8",
