--- test/mocha.opts
@@ -1,6 +1,0 @@
---require ts-node/register
---color
---recursive
---timeout 200000
---exit
-test/{,**/}*.spec.ts
