--- test/utils.ts
@@ -1,3 +1,0 @@
-export function pause(seconds: number): Promise<void> {
-  return new Promise((res) => setTimeout(res, 1000 * seconds));
-}
